Player and Orc fix
Fixed the Player and Orc assets in this branch. The Orc used sprites from a nested folder and needed to be reanimated. Select obj that has the animation before you can edit the animation. And needing to fix the scaling with 100->12 for each sprite. I copypasta'd the managers from another level to fix the player/cam ctrls